fix(app): rename loading to busy
